Split code and use a builder
