+++ .github/workflows/test.yml
@@ -3,6 +3,15 @@ name: Test
 on: push
 
 jobs:
+  no_diff:
+    name: Check no-diff from build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: zendesk/checkout@v2
+      - uses: zendesk/setup-ruby@v1.48.0
+        with:
+          bundler-cache: true
+      - run: ruby script/no-diff.rb
   test:
     name: rspec
     runs-on: ubuntu-latest
