+++ src/main.rb
@@ -1,343 +1,15 @@
-#!/usr/bin/ruby
-
-# This is a generated file; for the source, see 'src'.
-# Build with "ruby ./script/build.rb"
-
 #!/usr/bin/env ruby
 
 require 'set'
 require 'shellwords'
 require 'tempfile'
 
-require 'etc'
-require 'shellwords'
-require 'tempfile'
-require 'tmpdir'
-
-class MultiGitLsRunner
-  PREFIX_LENGTH = 6
-
-  def initialize(patterns)
-    # Each arg to the script will be "<output_file>:<pattern>"
-    @inputs = patterns.sort.uniq.each_with_index.map do |pattern, index|
-      [ pattern, "%0#{PREFIX_LENGTH}d" % index ]
-    end
-  end
-
-  def run
-    with_output_dir do
-      with_splitter_script do
-        run_xargs
-        outputs_by_pattern
-      end
-    end
-  end
-
-  private
-
-  attr_reader :inputs
-
-  def with_output_dir
-    Dir.mktmpdir do |tmpdir|
-      @output_dir = tmpdir
-      yield
-    end
-  end
-
-  def with_splitter_script
-    Tempfile.open do |script|
-      script.puts <<~SCRIPT
-        #!/bin/bash
-        umask 077
-        exec git ls-files -z --cached --ignored --exclude "${1:#{PREFIX_LENGTH + 1}}" \
-          > #{Shellwords.escape(@output_dir)}/${1:0:#{PREFIX_LENGTH}}
-      SCRIPT
-      script.chmod 0o700
-      script.close
-
-      @splitter = script.path
-      yield
-    end
-  end
-
-  def run_xargs
-    Tempfile.open do |input_file|
-      inputs.each do |pattern, output_file|
-        input_file.puts "#{output_file}:#{pattern}"
-      end
-      input_file.flush
-
-      system "xargs", "-n", "1", "-P", Etc.nprocessors.to_s, @splitter,
-             in: input_file.path
-      $?.success? or raise "xargs / git ls-files failed"
-    end
-  end
-
-  def outputs_by_pattern
-    inputs.map do |pattern, output_file|
-      file = File.join(@output_dir, output_file)
-      results = File.read(file).split("\0")
-      [pattern, results]
-    end.to_h
-  end
-end
-
-class Entry
-  def initialize(text:, file:, line_number:)
-    @text = text
-    @file = file
-    @line_number = line_number
-  end
-
-  attr_reader :text, :file, :line_number
-
-  def to_json(*args)
-    to_h.to_json(*args)
-  end
-
-  def to_h
-    { text: text, file: file, line_number: line_number }
-  end
-end
-
-class OwnerEntry < Entry
-  def initialize(text:, file:, line_number:, pattern:, owners:, indent:)
-    super(text: text, file: file, line_number: line_number)
-    @pattern = pattern
-    @owners = owners
-    @indent = indent
-  end
-
-  attr_reader :pattern, :owners, :indent
-
-  def to_h
-    super.merge(pattern: pattern, owners: owners, indent: indent)
-  end
-end
-
-require 'optparse'
-
-class GetOptions
-  def initialize(argv)
-    @debug = false
-    @show_json = false
-    @strict = false
-    @find_redundant_ignores = false
-    @who_owns = false
-    @files_owned = false
-    @check_unowned = false
-    @should_check_indent = true
-    @should_check_sorted = true
-    @should_check_valid_owners = true
-    read_options(argv)
-    validate
-  end
-
-  attr_reader :debug, :show_json, :strict, :find_redundant_ignores,
-              :who_owns, :files_owned,
-              :check_unowned, :should_check_indent, :should_check_sorted, :should_check_valid_owners,
-              :args
-
-  private
-
-  attr_writer :debug, :show_json, :strict, :find_redundant_ignores,
-              :who_owns, :files_owned,
-              :check_unowned, :should_check_indent, :should_check_sorted, :should_check_valid_owners,
-              :args
-
-  def read_options(argv)
-    parser = OptionParser.new do |opts|
-      opts.banner = "Usage: check-codeowners [options] [arguments]"
-
-      opts.on("-j", "--json", "Show results as JSON") do
-        self.show_json = true
-      end
-      opts.on("-s", "--strict", "Treat warnings as errors") do
-        self.strict = true
-      end
-      opts.on("-b", "--brute-force", "Find entries which do not match any files") do
-        self.find_redundant_ignores = true
-      end
-      opts.on("--no-check-indent", "Do not require equal indenting") do |value|
-        self.should_check_indent = value
-      end
-      opts.on("--no-check-sorted", "Do not require sorted entries") do |value|
-        self.should_check_sorted = value
-      end
-      opts.on("--no-check-valid-owners", "Do not check owners against VALIDOWNERS") do |value|
-        self.should_check_valid_owners = value
-      end
-      opts.on(
-        "--who-owns",
-        "Treat arguments as a list of files; show who owns each file, then exit." \
-        " If no arguments are given, all files are shown. Incompatible with" \
-        " --files-owned."
-      ) do
-        self.who_owns = true
-      end
-      opts.on(
-        "--files-owned",
-        "Treat arguments as a list of owners; show what files they own, then exit." \
-        " Incompatible with --who-owns."
-      ) do
-        self.files_owned = true
-      end
-      opts.on(
-        "--check-unowned",
-        "Checks if there are new files that are not owned"
-      ) do
-        self.check_unowned = true
-      end
-      opts.on("-d", "--debug", "Include debug output; implies --json") do
-        self.debug = true
-        self.show_json = true
-      end
-    end
-
-    argv = [*argv]
-    parser.parse!(argv)
-    @args = argv
-  end
-
-  def validate
-    if who_owns && files_owned
-      $stderr.puts "Invalid usage. Try check-codeowners --help"
-      exit 2
-    end
-
-    if !who_owns && !files_owned && args.any?
-      $stderr.puts "Invalid usage. Try check-codeowners --help"
-      exit 2
-    end
-  end
-end
-
-class Parsers
-  def parse_codeowners_file(path)
-    lines = begin
-              IO.readlines(path).map(&:chomp)
-            rescue Errno::ENOENT
-              []
-            end
-
-    owner_re = /\S+/
-
-    errors = []
-
-    entries = lines.each_with_index.map do |line, index|
-      base = { line_number: index + 1, text: line, file: path }
-
-      case line
-      when "", /^#/
-        # Could be used in the future to reconstruct the file
-        Entry.new(base)
-      when /^((\S+)\s+)(#{owner_re}( #{owner_re})*)$/
-        base.merge!(pattern: $2, indent: ($1).length, owners: $3.split(' '))
-        OwnerEntry.new(base)
-      else
-        Entry.new(base).tap do |entry|
-          errors << {
-            code: "unrecognised_line",
-            message: "Unrecognised line at #{entry.file}:#{entry.line_number}",
-            entry: entry
-          }
-        end
-      end
-    end
-
-    Struct.new(:entries, :errors).new(entries, errors)
-  end
-
-  def parse_ignore_file(path, options)
-    lines = begin
-              IO.readlines(path).map(&:chomp)
-            rescue Errno::ENOENT
-              []
-            end
-
-    patterns = Set.new
-    files = Set.new
-    previous_line = nil
-
-    errors = []
-
-    lines.each_with_index do |line, index|
-      next if line.empty? || line.start_with?("#")
-
-      if previous_line && line <= previous_line && options.should_check_sorted
-        errors << {
-          message: "Line is duplicated or out of sequence at #{path}:#{index + 1}",
-          code: "ignore_file_not_in_sequence",
-          file: path,
-          line: index + 1,
-        }
-      end
-
-      previous_line = line
-
-      if line.include?("*")
-        patterns << line
-      else
-        files << line
-      end
-    end
-
-    Struct.new(:patterns, :files, :errors).new(patterns, files, errors)
-  end
-
-  def parse_validowners(path)
-    begin
-      IO.readlines(path).map(&:chomp)
-    rescue Errno::ENOENT
-      nil
-    end
-  end
-end
-
-class Repository
-  CODEOWNERS_PATHS = [
-    "CODEOWNERS",
-    "docs/CODEOWNERS",
-    ".github/CODEOWNERS",
-  ].freeze
-
-  # The repository directory is always the current directory.
-  # Maybe that could be changed one day.
-
-  def codeowners_file
-    @codeowners_file ||= find_codeowners_file
-  end
-
-  def codeowners_ignore_file
-    @codeowners_ignore_file ||= codeowners_file + ".ignore"
-  end
-
-  def validowners_file
-    @validowners_file ||= codeowners_file.sub(/CODEOWNERS$/, "VALIDOWNERS")
-  end
-
-  def codeowners_entries
-    @codeowners_entries ||= Parsers.new.parse_codeowners_file(codeowners_file)
-  end
-
-  def ignore_entries(options)
-    @ignore_entries ||= Parsers.new.parse_ignore_file(codeowners_ignore_file, options)
-  end
-
-  def validowners_entries
-    @validowners_entries ||= Parsers.new.parse_validowners(validowners_file)
-  end
-
-  private
-
-  # https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#codeowners-file-location
-  def find_codeowners_file
-    CODEOWNERS_PATHS.find do |path|
-      File.exist?(path)
-    end || CODEOWNERS_PATHS.last
-  end
-end
-
+require_relative 'lib/check_codeowners/multi_git_ls_runner'
+require_relative 'lib/check_codeowners/entry'
+require_relative 'lib/check_codeowners/owner_entry'
+require_relative 'lib/check_codeowners/get_options'
+require_relative 'lib/check_codeowners/parsers'
+require_relative 'lib/check_codeowners/repository'
 
 def check_sorted(owner_entries)
   errors = []
